+++ cli/commands/check/interactive.go
@@ -30,6 +30,7 @@ type checkOpts struct {
 	Publish       string `survey:"publish"`
 	ProxyEntityID string `survey:"proxy-entity-id"`
 	Stdin         string `survey:"stdin"`
+	Timeout       string `survey:"timeout"`
 }
 
 func newCheckOpts() *checkOpts {
@@ -50,6 +51,7 @@ func (opts *checkOpts) withCheck(check *types.CheckConfig) {
 	opts.RuntimeAssets = strings.Join(check.RuntimeAssets, ",")
 	opts.ProxyEntityID = check.ProxyEntityID
 	opts.Stdin = stdinDefault
+	opts.Timeout = strconv.Itoa(int(check.Timeout))
 }
 
 func (opts *checkOpts) withFlags(flags *pflag.FlagSet) {
@@ -63,6 +65,7 @@ func (opts *checkOpts) withFlags(flags *pflag.FlagSet) {
 	opts.Publish = strconv.FormatBool(publishBool)
 	opts.ProxyEntityID, _ = flags.GetString("proxy-entity-id")
 	opts.Stdin, _ = flags.GetString("stdin")
+	opts.Timeout, _ = flags.GetString("timeout")
 
 	if org, _ := flags.GetString("organization"); org != "" {
 		opts.Org = org
@@ -137,6 +140,13 @@ func (opts *checkOpts) administerQuestionnaire(editing bool) error {
 			},
 		},
 		{
+			Name: "timeout",
+			Prompt: &survey.Input{
+				Message: "Timeout:",
+				Default: opts.Timeout,
+			},
+		},
+		{
 			Name: "subscriptions",
 			Prompt: &survey.Input{
 				Message: "Subscriptions:",
@@ -196,6 +206,7 @@ func (opts *checkOpts) administerQuestionnaire(editing bool) error {
 func (opts *checkOpts) Copy(check *types.CheckConfig) {
 	interval, _ := strconv.ParseUint(opts.Interval, 10, 32)
 	stdin, _ := strconv.ParseBool(opts.Stdin)
+	timeout, _ := strconv.ParseUint(opts.Timeout, 10, 32)
 
 	check.Name = opts.Name
 	check.Environment = opts.Env
@@ -209,4 +220,5 @@ func (opts *checkOpts) Copy(check *types.CheckConfig) {
 	check.Publish = opts.Publish == "true"
 	check.ProxyEntityID = opts.ProxyEntityID
 	check.Stdin = stdin
+	check.Timeout = uint32(timeout)
 }